+++ javascript/Quill/Pages/index.ts
@@ -1,3 +1,4 @@
 export * from './PageType';
 export * from './Page';
+export * from './ComposePage';
 // Add other modules as needed