+++ ChangeLog.md
@@ -1,3 +1,3 @@
 # Changes !
 
-- Moved Repository because I, me, paradoxg2 am a dumb ass!
+- Restructured all of the javascript